refactor all inline css to own files
