+++ src/App.tsx
@@ -1,6 +1,5 @@
 import UserProfilePage from './components/UserProfilePage';
 import NavigationBar from './components/NavigationBar.tsx';
-import Col from 'react-bootstrap/esm/Col';
 import { Navigate, Route, Routes } from 'react-router-dom';
 import StartPage from './components/StartPage';
 import LoginForm from './components/StartPage/LoginForm';
@@ -15,10 +14,10 @@ const App = () => {
           <Routes>
             <Route path="*" element={token && user ? <Navigate replace to={`/${user.username}`} />: <StartPage />} />
             <Route path="/login" element={<LoginForm/>} />
-            <Route path="/register" element={<Col className="ExploreView col-md-5 h-100 d-flex flex-column">Register</Col>} />
+            <Route path="/register" element={<div>Register</div>} />
             <Route path="/:userName" element={<UserProfilePage />} />
-            <Route path="/explore" element={<Col className="ExploreView col-md-5 h-100 d-flex flex-column">Explore</Col>} />
-            <Route path="/settings" element={<Col className="SettingsView col-md-5 h-100 d-flex flex-column">Settings</Col>} />
+            <Route path="/explore" element={<div>Explore</div>} />
+            <Route path="/settings" element={<div>Settings</div>} />
           </Routes>
           <div style={{'display': 'flex', 'flex': '1'}} />
     </div>
