+++ src/components/NavigationBar.tsx/index.css
@@ -1,0 +1,18 @@
+.NavLink {
+  display: inline-block;
+  width: fit-content;
+  text-decoration: none;
+  color: #191919;
+  font-size: large;
+}
+
+.NavLink:hover {
+  color: black;
+}
+
+.NavBar {
+  display: flex; 
+  flex: 1;
+  flex-direction: column; 
+  align-items: flex-end;
+}
