+++ src/components/Post/Post/PostButtons.tsx
@@ -1,8 +1,9 @@
 import Button from 'react-bootstrap/esm/Button';
+import './index.css';
 
 const PostButtons = () => {
   return (
-    <div style={{'display': 'flex'}}>
+    <div className="PostButtons">
       <Button variant="dark">Reply</Button>
       <Button variant="dark">Like</Button>
     </div>
