+++ src/components/Post/Post/index.css
@@ -1,0 +1,16 @@
+
+.PostView {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  border: 1px solid;
+}
+
+
+.PostButtons {
+  display: flex;
+}
+
+.PostStats {
+  display: flex;
+}
