+++ src/components/Post/PostForm/index.css
@@ -1,4 +1,8 @@
-
+.PostForm {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+}
 
 
 .PostButton {
