+++ src/components/Post/PostList/index.css
@@ -1,0 +1,5 @@
+.PostList {
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+}
