+++ src/components/StartPage/index.css
@@ -1,10 +1,4 @@
 
-form {
-  display: inline-block; 
-  text-align: center;
-  position: relative;
-  width: 80%;
-}
 
 .StartPage {
   background-color: lightblue;
@@ -42,6 +36,7 @@ form {
   color: red;
 }
 
-#row {
-  min-height: 30%;
+.LogRegButtons {
+  display: flex;
+  padding-top: 3rem;
 }
