+++ src/components/UserProfilePage/index.css
@@ -1,0 +1,11 @@
+
+
+.UserProfilePage {
+  display: flex; 
+  flex: 1;
+  flex-direction: column;
+}
+
+.UserDetails {
+  border: 1px solid;
+}
